backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain-google-genai==0.0.5
google-auth==2.23.4
google-api-python-client==2.108.0
gspread==5.12.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0 